// 1xbts-web/src/lib/proto/events/v1/msc.ts
export enum OtaspFeatureId {
  OTASP_FEATURE_ID_NAM_DOWNLOAD = 0,
  OTASP_FEATURE_ID_KEY_EXCHANGE = 1,
  OTASP_FEATURE_ID_SSPR = 2,
  OTASP_FEATURE_ID_SERVICE_PROGRAMMING_LOCK = 3,
  OTASP_FEATURE_ID_OTASP = 4,
  OTASP_FEATURE_ID_PUZL = 5,
  OTASP_FEATURE_ID_PACKET_DATA_3GPD = 6,
  OTASP_FEATURE_ID_SECURE_MODE = 7,
  OTASP_FEATURE_ID_MMD = 8,
  OTASP_FEATURE_ID_SYSTEM_TAG_DOWNLOAD = 9,
  OTASP_FEATURE_ID_MMS = 10,
  OTASP_FEATURE_ID_MMSS = 11,
  UNRECOGNIZED = -1,
}